pyproject.toml
[tool.poetry]
name = "manifestservice"
version = "0.5.2"
description = "Handles reading from and writing to a user's s3 folder containing their manifests"
authors = ["CTDS UChicago <cdis@uchicago.edu>"]
license = "Apache-2.0"
repository = "https://github.com/uc-cdis/manifestservice"
include = [
  "NOTICE",
]

[tool.poetry.dependencies]
python = ">=3.9, <4.0"
Flask = "^2.3.2"
authutils = "<7"
boto3 = "^1.17.48"
cryptography = ">=42.0.5"
requests = ">=2.32.0"
gunicorn = ">=22.0.0"
importlib-metadata = "^4.8.1"

[tool.poetry.group.dev.dependencies]
pytest = "^6.2.3"
pytest-mock = "^3.5.1"
pytest-flask = "^1.2.0"
unittest2 = "^1.1.0"
codacy-coverage = "^1.3.11"
pyyaml = "^6.0.1"
flasgger = "^0.9.7.1"
isort = ">=5.12.0"
black = ">=23.10.0"
pylint = ">=3.0.1"

[build-system]
requires = ["poetry>=1.1.15"]
build-backend = "poetry.core.masonry.api"
